+++ packages/tests/all/jab Waiter cancel 3.ja.ts
@@ -2,13 +2,12 @@ import { TestProvision } from "^jarun";
 
 import { getWaiter } from "../_fixture";
 
-// await twice.
+// after timeout cancel is a noop, because the waiter threw.
 
 export default (prov: TestProvision) => {
   const waiter = getWaiter(prov);
 
-  prov.await(waiter.await("done", 1));
-  prov.await(waiter.await("done", 1));
-
-  waiter.set("done");
+  return waiter.await("done", 1).catch(() => {
+    waiter.cancel();
+  });
 };