+++ packages/tests/all/jab Waiter set 1.ja.ts
@@ -2,13 +2,14 @@ import { TestProvision } from "^jarun";
 
 import { getWaiter } from "../_fixture";
 
-// await twice.
+// set state when awaiting
 
 export default (prov: TestProvision) => {
   const waiter = getWaiter(prov);
 
-  prov.await(waiter.await("done", 1));
-  prov.await(waiter.await("done", 1));
+  const prom = waiter.await("done");
 
   waiter.set("done");
+
+  return prom;
 };