+++ packages/tests/all/jab Waiter set 2.ja.ts
@@ -2,13 +2,14 @@ import { TestProvision } from "^jarun";
 
 import { getWaiter } from "../_fixture";
 
-// await twice.
+// set state when awaiting for other state
 
 export default (prov: TestProvision) => {
   const waiter = getWaiter(prov);
 
-  prov.await(waiter.await("done", 1));
-  prov.await(waiter.await("done", 1));
+  const prom = waiter.await("done", 1);
 
-  waiter.set("done");
+  waiter.set("ready"); //has no effect
+
+  return prom; //should timeout
 };